+++ lib/repositories/music_api/models/song.dart
@@ -16,12 +16,16 @@ class Song {
   final String audioUrl;
 
   factory Song.fromJson(Map<String, dynamic> json) {
+    final originalCoverUrl = json['artworkUrl100'] as String? ?? '';
+    final coverUrl = originalCoverUrl.isEmpty ? ''
+        : originalCoverUrl.replaceAll('100x100', '400x400');
+
     return Song(
       id: json['trackId'] as int? ?? 0,
       title: json['trackName'] as String? ?? 'Unknown Track',
       artistName: json['artistName'] as String? ?? 'Unknown Artist',
       albumName: json['collectionName'] as String? ?? 'Unknown Album',
-      coverUrl: json['artworkUrl100'] as String? ?? '',
+      coverUrl: coverUrl,
       audioUrl: json['previewUrl'] as String? ?? '',
     );
   }