+++ lib/router/router.dart
@@ -1,7 +1,9 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 import '../app/di.dart';
+import '../features/game_flow/logic/game_provider.dart';
 import '../features/game_flow/view/game/view/game_screen.dart';
 import '../features/game_flow/view/game_setup/view/game_setup_screen.dart';
 import '../features/game_flow/view/game_summary/view/game_summary_screen.dart';
@@ -19,6 +21,7 @@ final router = GoRouter(
   observers: [
     TalkerRouteObserver(G<Talker>()),
   ],
+  initialLocation: AppRoute.root,
   errorBuilder: (_, __) => const NotFoundScreen(),
   routes: [
     GoRoute(
@@ -32,6 +35,15 @@ final router = GoRouter(
     GoRoute(
       path: AppRoute.game,
       builder: (_, __) => const GameScreen(),
+      redirect: (context, state) {
+        final container = ProviderScope.containerOf(context);
+        final questions = container.read(gameProvider).questions;
+
+        if (questions.isEmpty) {
+          return AppRoute.gameSetup;
+        }
+        return null;
+      },
       routes: [
         GoRoute(
           path: AppRoute.gameSummary,